// src/app/Model/painel-usuario.ts
export class PainelUsuario {
    
    tipoPublicacao?: string;
    nomeUser?: string;
    userId?: string;
    areaAtuacao?: string;
    profissao?: string;
    tempoExperiencia?: string;
    valorHora?: number;
    descricao?: string;
    dataPublicacao?: number;
    idUsuariologado?: string;
    idPublicacao?: string;
    status?: string;
    idMovimentacao?: number;
    horaInicio?: number;

}